Displaying fee with 2 decimal places

+++ app/models/forms/application/detail.rb
@@ -8,7 +8,7 @@ module Forms
 
       # rubocop:disable MethodLength
       def self.permitted_attributes
-        { fee: Integer,
+        { fee: Decimal,
           jurisdiction_id: Integer,
           date_received: Date,
           day_date_received: Integer,
